% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/census.helper.api.R
\name{census.helper.api}
\alias{census.helper.api}
\title{Census helper function.}
\usage{
census.helper.api(key, voters, states = "all", geo = "tract", demo = F)
}
\arguments{
\item{key}{A required character object. Must contain user's Census API
key, which can be requested at http://api.census.gov/data/key_signup.html.}

\item{voters}{An object of class \code{data.frame}. Must contain a field 
named 'county', 'tract', and/or 'block' that specifies geolocation.}

\item{states}{A character vector specifying which states to extract 
Census data for, e.g. c("NJ", "NY"). Default is "all", which extracts 
Census data for all states contained in user-input data.}

\item{geo}{A character object specifying what aggregation level to use. 
Use "county", "tract", or "block". Default is "tract". 
Warning: block takes very long.}

\item{demo}{A TRUE/FALSE object indicating whether to condition on 
demographics (i.e., age and sex) or not. If TRUE, function will return 
Pr(Geolocation, Age, Sex | Race). If FALSE, function wil return 
Pr(Geolocation | Race). Default is FALSE.}
}
\value{
Output will be an object of class \code{data.frame}. It will 
 consist of the original user-input data with additional columns of 
 Census data.
}
\description{
\code{census.helper.api} links user-input dataset with Census data.
}
\details{
This function allows users to link their geocoded dataset (e.g., voter file) 
with U.S. Census 2010 data. The function extracts Census Summary File data 
at the tract or block level using the 'UScensus2010' package. Census data 
calculated are Pr(Geolocation | Race) where geolocation is tract or block.
}
\examples{
\dontshow{data(voters)}
\dontrun{census.helper.api(key = "...", voters = voters, states = "nj", geo = "block")}
\dontrun{census.helper.api(key = "...", voters = voters, states = "all", geo = "tract", 
demo = TRUE)}

}
\references{
Relies on getCensusApi, getCensusApi2, and vecToChunk functions authored by Nicholas Nagle, 
available at: http://rstudio-pubs-static.s3.amazonaws.com/19337_2e7f827190514c569ea136db788ce850.html.
}

