% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/census.helper.R
\name{census.helper}
\alias{census.helper}
\title{Census helper function.}
\usage{
census.helper(voters, states, geo)
}
\arguments{
\item{voters}{An object of class \code{data.frame}. Must contain a field 
named 'tract' or 'block' that specifies geolocation.}

\item{states}{A character vector specifying which states to extract 
Census data for, e.g. c("NJ", "NY"). May also be "all" to extract 
Census data for all states contained in user-input data.}

\item{geo}{A character object specifying what aggregation level to use. 
Use "trt" for tract or "blk" for block. Warning: block takes very long.}
}
\value{
Output will be an object of class \code{data.frame}. It will 
 consist of the original user-input data with additional columns of 
 Census data.
}
\description{
\code{census.helper} links user-input dataset with Census data.
}
\details{
This function allows users to link their geocoded dataset (e.g., voter file) 
with U.S. Census 2010 data. The function extracts Census data at the tract 
or block level using the 'UScensus2010' package. Census data calculated are 
Pr(geolocation | race) where geolocation is tract or block.
}
\examples{
census.helper(NJ.voter.file, "nj", "blk")
census.helper(voters.all, c("nj", "ny"), "trt")

}

