% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/census_geo_api_joint.R
\name{census_geo_api_joint}
\alias{census_geo_api_joint}
\title{Census data download function, returning joint frequencies}
\usage{
census_geo_api_joint(
  key,
  state,
  geo = "tract",
  age = FALSE,
  sex = FALSE,
  retry = 0
)
}
\arguments{
\item{key}{A required character object. Must contain user's Census API key, which can be requested here.}

\item{state}{A required character object specifying which state to extract Census data for, e.g., "NJ".}

\item{geo}{A character object specifying what aggregation level to use. Use "county", "tract", "block",
or "place". Default is "tract". Warning: extracting block-level data takes very long.}

\item{age}{A TRUE/FALSE object indicating whether to condition on age or not. If FALSE (default), function
will return Pr(Geolocation, Race). If TRUE, function will return Pr(Geolocation, Age, Race). 
If sex is also TRUE, function will return Pr(Geolocation, Age, Sex, Race).}

\item{sex}{A TRUE/FALSE object indicating whether to condition on sex or not. If FALSE (default), 
function will return Pr(Geolocation, Race). If TRUE, function will return 
Pr(Geolocation, Sex, Race). If age is also TRUE, function will return 
Pr(Geolocation, Age, Sex, Race).}

\item{retry}{The number of retries at the census website if network interruption occurs.}
}
\value{
Output will be an object of class list, indexed by state names. 
        It will consist of the original user-input data with additional columns of Census geographic data, 
        containing joint frequency distributions according to the census.
}
\description{
This function allows users to download U.S. Census 2010 geographic data,
at either the county, tract, block, or place level, for a particular state.
It differs from \code{census_geo_api} in that the this function returns a 
full joint frequency distribution, whereas that function returns Pr(. | Race) 
conditionals.
}
