% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/census_geo_api_zcta.R
\name{census_geo_api_zcta}
\alias{census_geo_api_zcta}
\title{Census download function for state-ZCTA-level data}
\usage{
census_geo_api_zcta(
  state,
  ...,
  age = FALSE,
  sex = FALSE,
  year = c("2020", "2010", "2000"),
  retry = 3,
  key = Sys.getenv("CENSUS_API_KEY")
)
}
\arguments{
\item{state}{A required character object specifying which state to extract Census data for,
e.g., \code{"NJ"}.}

\item{...}{These dots are for future extensions and must be empty.}

\item{age}{A \code{TRUE}/\code{FALSE} object indicating whether to condition on
age or not. If \code{FALSE} (default), function will return Pr(Geolocation | Race).
If \code{TRUE}, function will return Pr(Geolocation, Age | Race).
If \code{\var{sex}} is also \code{TRUE}, function will return Pr(Geolocation, Age, Sex | Race).}

\item{sex}{A \code{TRUE}/\code{FALSE} object indicating whether to condition on
sex or not. If \code{FALSE} (default), function will return Pr(Geolocation | Race).
If \code{TRUE}, function will return Pr(Geolocation, Sex | Race).
If \code{\var{age}} is also \code{TRUE}, function will return Pr(Geolocation, Age, Sex | Race).}

\item{year}{A character object specifying the year of U.S. Census data to be downloaded.
Use \code{"2010"}, or \code{"2020"}. Default is \code{"2020"}.
Warning: 2020 U.S. Census data is downloaded only when \code{\var{age}} and
\code{\var{sex}} are both \code{FALSE}.}

\item{retry}{The number of retries at the census website if network interruption occurs.}

\item{key}{A required character object. Must contain user's Census API
key, which can be requested \href{https://api.census.gov/data/key_signup.html}{here}.}
}
\value{
A \link{data.frame} with columns
\code{state}, \code{zcta}, \code{r_whi}, \code{r_bla}, \code{r_his}, \code{r_asi}, and \code{r_oth}.
}
\description{
Census download function for state-ZCTA-level data
}
\examples{
\dontshow{if (nzchar(Sys.getenv("CENSUS_API_KEY"))) (if (getRversion() >= "3.4") withAutoprint else force)(\{ # examplesIf}
\dontrun{census_geo_api_zcta(state = c("DE", "NJ")}
\dontrun{census_geo_api_zcta(state = "FL", age = TRUE, sex = TRUE)}
\dontrun{census_geo_api_zcta(state = "MA", age = FALSE, sex = FALSE, year = "2020")}
\dontshow{\}) # examplesIf}
}
\keyword{internal}
