% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/census_helper.R
\name{census_helper}
\alias{census_helper}
\title{Census helper function.}
\usage{
census_helper(key, voter.file, states = "all", geo = "tract",
  demo = FALSE, census.data = NA)
}
\arguments{
\item{key}{A required character object. Must contain user's Census API
key, which can be requested \href{http://api.census.gov/data/key_signup.html}{here}.}

\item{voter.file}{An object of class \code{data.frame}. Must contain field(s) 
named \code{\var{county}}, \code{\var{tract}}, and/or \code{\var{block}} 
specifying geolocation. These should be character variables that match up with 
U.S. Census categories. County should be three characters (e.g., "031" not "31"), 
tract should be six characters, and block should be four characters.}

\item{states}{A character vector specifying which states to extract 
Census data for, e.g. \code{c("NJ", "NY")}. Default is \code{"all"}, which extracts 
Census data for all states contained in user-input data.}

\item{geo}{A character object specifying what aggregation level to use. 
Use \code{"county"}, \code{"tract"}, or \code{"block"}. Default is \code{"tract"}. 
Warning: extracting block-level data takes very long.}

\item{demo}{A \code{TRUE}/\code{FALSE} object indicating whether to condition on 
demographics (i.e., age and sex) or not. If \code{TRUE}, function will return 
Pr(Geolocation, Age, Sex | Race). If \code{FALSE}, function wil return 
Pr(Geolocation | Race). Default is \code{FALSE}.}

\item{census.data}{A optional census object of class \code{list} containing 
pre-saved Census geographic data. Can be created using \code{get_census_data} function.
If \code{\var{census.data}} is provided, its \code{\var{demo}} element must have the 
same value as the \code{\var{demo}} option specified in this function.
If missing, function will retrive the census data via Census API.}
}
\value{
Output will be an object of class \code{data.frame}. It will 
 consist of the original user-input data with additional columns of 
 Census data.
}
\description{
\code{census_helper} links user-input dataset with Census geographic data.
}
\details{
This function allows users to link their geocoded dataset (e.g., voter file) 
with U.S. Census 2010 data. The function extracts Census Summary File data 
at the county, tract, or block level using the 'UScensus2010' package. Census data 
calculated are Pr(Geolocation | Race) where geolocation is county, tract, or block.
}
\examples{
\dontshow{data(voters)}
\dontrun{census_helper(key = "...", voter.file = voters, states = "nj", geo = "block")}
\dontrun{census_helper(key = "...", voter.file = voters, states = "all", geo = "tract", demo = TRUE)}

}

