% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/census_helper_local.R
\name{census_helper_local}
\alias{census_helper_local}
\title{Census helper function.}
\usage{
census_helper_local(voter.file, states = "all", geo = "tract",
  demo = FALSE, census.data = NA)
}
\arguments{
\item{voter.file}{An object of class \code{data.frame}. Must contain field(s) 
named \code{\var{county}}, \code{\var{tract}}, and/or \code{\var{block}} 
specifying geolocation. These should be character variables that match up with 
U.S. Census categories. County should be three characters (e.g., "031" not "31"), 
tract should be six characters, and block should be four characters.}

\item{states}{A state to use Census data for, e.g. \code{c("NJ", "NY")}. 
Default is \code{"NA"}.}

\item{geo}{A character object specifying what aggregation level to use. 
Use \code{"county"}, \code{"tract"}, or \code{"block"}. Default is \code{"tract"}. 
Warning: extracting block-level data takes very long.}

\item{demo}{A \code{TRUE}/\code{FALSE} object indicating whether to condition on 
demographics (i.e., age and sex) or not. If \code{TRUE}, function will return 
Pr(Geolocation, Age, Sex | Race). If \code{FALSE}, function wil return 
Pr(Geolocation | Race). Default is \code{FALSE}.}

\item{census.data}{An optional census object holding Census data that is already 
provided. If missing, function will retrieve the Census geographic data via API.}
}
\value{
Output will be an object of class \code{data.frame}. It will 
 consist of the original user-input data with additional columns of 
 Census data.
}
\description{
\code{census_helper_local} links user-input dataset with Census data.
}
\details{
This function allows users to link their geocoded dataset (e.g., voter file) 
with U.S. Census 2010 data. The function extracts Census Summary File data 
at the county, tract, or block level. Function returns Pr(Geolocation | Race) 
where geolocation is county, tract, or block.
}
\examples{
\dontshow{data(voters)}
\dontrun{census_helper_local(voter.file = voters, states = "nj", geo = "block", census.data = x)}
\dontrun{census_helper_local(voter.file = voters, states = "all", geo = "tract", demo = TRUE, census.data = x)}

}
\references{
Relies on get_census_api, get_census_api_2, and vec_to_chunk functions authored by Nicholas Nagle, 
available \href{http://rstudio-pubs-static.s3.amazonaws.com/19337_2e7f827190514c569ea136db788ce850.html}{here}.
}

