% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/census_helper_v2.R
\name{census_helper_new}
\alias{census_helper_new}
\title{Census helper function.}
\usage{
census_helper_new(
  key,
  voter.file,
  states = "all",
  geo = "tract",
  year = "2010",
  census.data = NA,
  retry = 0
)
}
\arguments{
\item{key}{A required character object. Must contain user's Census API
key, which can be requested \href{https://api.census.gov/data/key_signup.html}{here}.}

\item{voter.file}{An object of class \code{data.frame}. Must contain field(s) named 
\code{\var{county}}, \code{\var{tract}}, \code{\var{block}}, and/or \code{\var{place}}
specifying geolocation. These should be character variables that match up with 
U.S. Census categories. County should be three characters (e.g., "031" not "31"), 
tract should be six characters, and block should be four characters. 
Place should be five characters if it is included.}

\item{states}{A character vector specifying which states to extract 
Census data for, e.g. \code{c("NJ", "NY")}. Default is \code{"all"}, which extracts 
Census data for all states contained in user-input data.}

\item{geo}{A character object specifying what aggregation level to use. 
Use \code{"county"}, \code{"tract"}, \code{"block"}, or \code{"place"}. 
Default is \code{"tract"}. Warning: extracting block-level data takes very long.}

\item{year}{A character object specifying the year of U.S. Census data to be downloaded.
Use \code{"2010"}, or \code{"2020"}. Default is \code{"2010"}.}

\item{census.data}{A optional census object of class \code{list} containing 
pre-saved Census geographic data. Can be created using \code{get_census_data} function.
If \code{\var{census.data}} is provided, the \code{\var{year}} element must 
have the same value as the \code{\var{year}} option specified in this function 
(i.e., \code{"2010"} in both or \code{"2020"} in both). 
If \code{\var{census.data}} is provided, the \code{\var{age}} and the \code{\var{sex}} 
elements must be \code{FALSE}. This corresponds to the defaults of \code{census_geo_api}.
If \code{\var{census.data}} is missing, Census geographic data will be obtained via Census API.}

\item{retry}{The number of retries at the census website if network interruption occurs.}
}
\value{
Output will be an object of class \code{data.frame}. It will 
 consist of the original user-input data with additional columns of 
 Census data.
}
\description{
\code{census_helper_new} links user-input dataset with Census geographic data.
}
\details{
This function allows users to link their geocoded dataset (e.g., voter file) 
with U.S. Census data (2010 or 2020). The function extracts Census Summary File data 
at the county, tract, block, or place level. Census data calculated are 
Pr(Geolocation | Race) where geolocation is county, tract, block, or place.
}
\examples{
\dontshow{data(voters)}
\dontrun{census_helper_new(key = "...", voter.file = voters, states = "nj", geo = "block")}
\dontrun{census_helper_new(key = "...", voter.file = voters, states = "all", geo = "tract")}
\dontrun{census_helper_new(key = "...", voter.file = voters, states = "all", geo = "place", year = "2020")}

}
