% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/name_preproc.R
\name{.name_preproc}
\alias{.name_preproc}
\title{Pre-process vector of names to match census style. Internal function}
\usage{
.name_preproc(voter_names, target_names)
}
\arguments{
\item{voter_names}{Character vector to be pre-processed.}

\item{target_names}{Character vector of census names to be matched.}
}
\value{
A character vector of pre-processed named
}
\description{
Pre-process vector of names to match census style. Internal function
}
\keyword{internal}
