% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ebisg_utils.R
\name{ebisg_data_preflight}
\alias{ebisg_data_preflight}
\title{Download eBISG model weights via piggyback}
\usage{
ebisg_data_preflight(cfg)
}
\arguments{
\item{cfg}{A model config list from \code{resolve_ebisg_model()}.}
}
\value{
The destination directory path.
}
\description{
Download eBISG model weights via piggyback
}
\keyword{internal}
