% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ebisg_utils.R
\name{ebisg_predict_mlp}
\alias{ebisg_predict_mlp}
\title{Run MLP inference on embeddings}
\usage{
ebisg_predict_mlp(embeddings, model_path)
}
\arguments{
\item{embeddings}{Numeric matrix (N x dim) of name embeddings.}

\item{model_path}{Path to the .pt checkpoint file.}
}
\value{
A matrix of dimensions (N, 6) with columns
(whi, bla, his, asi, aian, oth).
}
\description{
Run MLP inference on embeddings
}
\keyword{internal}
