% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lbisg.R
\name{ensure_lbisg_python}
\alias{ensure_lbisg_python}
\title{Check that Python dependencies for lBISG are available}
\usage{
ensure_lbisg_python()
}
\description{
Check that Python dependencies for lBISG are available
}
\keyword{internal}
