% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\name{firstNameDict}
\alias{firstNameDict}
\title{Dictionary of First Names}
\format{
A data frame with 1,043,742 rows and 6 columns:
\describe{
  \item{first_name}{Voter identifier (numeric)}
  \item{p_whi_first}{Probability of white}
  \item{p_bla_first}{Probability of black}
  \item{p_his_first}{Probability of Hispanic}
  \item{p_asi_first}{Probability of asian},
  \item{p_oth_first}{Probability of other}
  }
}
\description{
An example dataset containing voter file information.
}
\examples{
data(firstNameDict)
}
\keyword{datasets}
