% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getCensusApi.R
\name{getCensusApi}
\alias{getCensusApi}
\title{get Census data via the public API: loop through variables if needed}
\usage{
getCensusApi(data_url, key, vars, region, numeric = TRUE)
}
\description{
get Census data via the public API: loop through variables if needed
}
\examples{
getCensusApi(data_url = "http://api.census.gov/data/2010/sf1?", key = "", region = "")

}

