% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getCensusApi2.R
\name{getCensusApi2}
\alias{getCensusApi2}
\title{getCensusApi2}
\usage{
getCensusApi2(data_url, key, get, region, numeric = TRUE)
}
\description{
getCensusApi2
}
\examples{
getCensusApi2(data_url = "http://api.census.gov/data/2010/sf1?", key = "", region = "")

}

