% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\name{lastNameDict}
\alias{lastNameDict}
\title{Dictionary of Last Names}
\format{
A data frame with 1,502,541 rows and 6 columns:
\describe{
  \item{last_name}{Voter identifier (numeric)}
  \item{p_whi_last}{Probability of white}
  \item{p_bla_last}{Probability of black}
  \item{p_his_last}{Probability of Hispanic}
  \item{p_asi_last}{Probability of asian},
  \item{p_oth_last}{Probability of other}
  }
}
\description{
An example dataset containing voter file information.
}
\examples{
data(lastNameDict)
}
\keyword{datasets}
