% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ebisg_utils.R
\name{map_6class_to_5class}
\alias{map_6class_to_5class}
\title{Map 6-class eBISG output to wru's 5-class system}
\usage{
map_6class_to_5class(probs_6)
}
\arguments{
\item{probs_6}{A matrix with 6 columns.}
}
\value{
A matrix with 5 columns named c_whi, c_bla, c_his, c_asi, c_oth.
}
\description{
The MLP outputs probabilities for 6 categories:
(white, black, hispanic, asian, aian, other).
wru uses 5 categories: (whi, bla, his, asi, oth).
This function sums aian and other into oth.
}
\keyword{internal}
