% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\name{middleNameDict}
\alias{middleNameDict}
\title{Dictionary of Middle Names}
\format{
A data frame with 1,182,133 rows and 6 columns:
\describe{
  \item{middle_name}{Voter identifier (numeric)}
  \item{p_whi_middle}{Probability of white}
  \item{p_bla_middle}{Probability of black}
  \item{p_his_middle}{Probability of Hispanic}
  \item{p_asi_middle}{Probability of asian},
  \item{p_oth_middle}{Probability of other}
  }
}
\description{
An example dataset containing voter file information.
}
\examples{
data(middleNameDict)
}
\keyword{datasets}
