\name{names}
\alias{names}
\title{Dataset containing Census Surname List.}
\description{
A dataset containing Census Surname List, which was released by the U.S. Census Bureau in 2000. Variables are as follows:
  \itemize{
  \item surname
  \item p_whi (i.e., Pr(White | Surname)
  \item p_bla (i.e., Pr(Black | Surname)
  \item p_his (i.e., Pr(Hispanic/Latino | Surname)
  \item p_asi (i.e., Pr(Asian | Surname)
  \item p_oth (i.e., Pr(Other | Surname)
  }
}
\format{
A data frame with 151,671 rows and 6 variables.
}
