% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predict_race_hwru.R
\name{predict_race_hwru}
\alias{predict_race_hwru}
\title{Race prediction function, for hierarchical WRU model}
\usage{
predict_race_hwru(
  voter.file,
  namesToUse,
  census.geo,
  race.init = NULL,
  ...,
  control = NULL
)
}
\arguments{
\item{voter.file}{An object of class data.frame. Must contain a row for each individual being predicted, 
as well as fields named "surname", and/or"first", and/or "middle", containing each 
individual's corresponding names. It must also contain a field named state, which 
contains the two-character lower-case abbreviation for each individual's state of residence (e.g., "nj" for New Jersey).
voter.file must also contain a field with county codes and any additional levels (i.e. tract, block) for lower levels of aggregation.
These fields should contain character strings matching U.S. Census categories. County is three
characters (e.g., "031" not "31"), tract is six characters, and block is four characters. Place is
five characters. See \code{\link{predict_race}} for other optional fields.}

\item{namesToUse}{Character string Must be one of 'last', 'last, first', or 'last, first, middle'.}

\item{census.geo}{Required character vector. One of "county", "tract", "block" or "place". See \code{\link{predict_race}}.}

\item{race.init}{Initial race for each observation in voter.file. Must be an integer, with
0=white, 1=black, 2=hispanic, 3=asian, and 4=other. Defaults to values obtained from \code{race_predict}.}

\item{...}{Other arguments passed to \code{\link{predict_race}}.}

\item{control}{List of control arguments, including 
\itemize{
 \item{iter}{ Number of MCMC iterations. Defaults to 1000.}
 \item{burnin}{ Number of iterations discarded as burnin. Defaults to half of \code{iter}.}
 \item{verbose}{ Print progress information. Defaults to \code{TRUE}.}
 \item{progress_bar}{ If \code{verbose=TRUE}, should a progress bar be used to track progress? If \code{TRUE}, must be handled by \code{progressr::handlers()},
                      as per \code{progressr}'s design choice.}
 \item{seed}{ RNG seed. If \code{NULL}, a seed is generated and returned for reproducibility.}
}}
}
\value{
A named list:
\itemize{
\item{race_by_record}{ A copy of \code{voter.file}, with additional columns of predicted
                      race probabilities, names \code{pred.<race>}. }
 \item{seed}{ RNG seed used.}                     
}
}
\description{
Sample individual-level race/ethnicity from a hirarchical Bayesian Dirichlet-multinomial model.
}
