% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/State.FIPS.R
\docType{data}
\name{State.FIPS}
\alias{State.FIPS}
\title{State's FIPS.}
\format{A data frame with 55 rows and 2 variables:
\describe{
  \item{State}{State}
  \item{FIPS}{FIPS}
  #' }}
\usage{
State.FIPS
}
\description{
List of States and their FIPS.
}
\examples{
data(State.FIPS)
}
\keyword{datasets}
