% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/state_fips.R
\docType{data}
\name{state_fips}
\alias{state_fips}
\title{Dataset with FIPS codes for US states}
\format{
A tibble with 57 rows and 3 columns:
\describe{
\item{\code{state}}{Two-letter postal abbreviation}
\item{\code{state_code}}{Two-digit FIPS code}
\item{\code{state_name}}{English name}
}
}
\source{
Derived from \code{\link[tidycensus:fips_codes]{tidycensus::fips_codes()}}
}
\usage{
state_fips
}
\description{
Dataset including FIPS codes and postal abbreviations for each U.S. state,
district, and territory.
}
\keyword{datasets}
