% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/surnames2000.R
\docType{data}
\name{surnames2000}
\alias{surnames2000}
\title{Census Surname List (2000).}
\format{A data frame with 157,728 rows and 6 variables:
\describe{
  \item{surname}{Surname}
  \item{p_whi}{Pr(White | Surname)}
  \item{p_bla}{Pr(Black | Surname)}
  \item{p_his}{Pr(Hispanic/Latino | Surname)}
  \item{p_asi}{Pr(Asian/Pacific Islander | Surname)}
  \item{p_oth}{Pr(Other | Surname)}
  #' }}
\usage{
surnames2000
}
\description{
Census Surname List from 2000 with race/ethnicity probabilities by surname.
}
\examples{
data(surnames2000)
}
\keyword{datasets}

