% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/surnames2010.R
\docType{data}
\name{surnames2010}
\alias{surnames2010}
\title{Census Surname List (2010).}
\format{A data frame with 167,613 rows and 6 variables:
\describe{
  \item{surname}{Surname}
  \item{p_whi}{Pr(White | Surname)}
  \item{p_bla}{Pr(Black | Surname)}
  \item{p_his}{Pr(Hispanic/Latino | Surname)}
  \item{p_asi}{Pr(Asian/Pacific Islander | Surname)}
  \item{p_oth}{Pr(Other | Surname)}
  #' }}
\usage{
surnames2010
}
\description{
Census Surname List from 2010 with race/ethnicity probabilities by surname.
}
\examples{
data(surnames)
}
\keyword{datasets}
