\name{voters}
\alias{voters}
\title{Dataset containing example voter file.}
\description{
A dataset containing an example voter file with 10 voters. Variables are as follows:
  \itemize{
  \item VoterID
  \item surname
  \item state
  \item CD (i.e., Congressional District)
  \item county
  \item tract
  \item block
  \item precinct
  \item age
  \item female
  \item party
  }
}
\format{
A data frame with 10 rows and 11 variables.
}
