% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/state_fips.R
\name{as_fips_code}
\alias{as_fips_code}
\title{Get the standardized FIPS code associated with a state}
\usage{
as_fips_code(x)
}
\arguments{
\item{x}{A \link{numeric} or \link{character} vector of state names,
postal abbreviations, or FIPS codes.
Matches for state names and abbreviations are not case sensitive.
FIPS codes may be matched from numeric or character vectors,
with or without leading zeroes.}
}
\value{
A \link{character} vector of two-digit FIPS codes.
One-digit FIPS codes are prefixed with a leading zero,
e.g., \code{"06"} for California.
}
\description{
Get the standardized FIPS code associated with a state
}
\examples{
as_fips_code("california")

# Character vector matches ignore case
as_fips_code(c("DC", "Md", "va"))

# Note that `3` and `7` are standardized to `NA`,
# because no state is assigned those FIPS codes
as_fips_code(1:10)

# You can even mix methods in the same vector
as_fips_code(c("utah", "NM", 8, "04"))

}
\keyword{internal}
