% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/census_data_preflight.R
\name{census_data_preflight}
\alias{census_data_preflight}
\title{Preflight census data}
\usage{
census_data_preflight(census.data, census.geo, year)
}
\arguments{
\item{census.data}{See documentation in \code{race_predict}.}

\item{census.geo}{See documentation in \code{race_predict}.}

\item{year}{See documentation in \code{race_predict}.}
}
\description{
Preflight census data
}
