% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/co_cluster.R
\name{co_cluster}
\alias{co_cluster}
\title{Name-Race Co-clustering Using Keynames}
\usage{
co_cluster(
  voter.file,
  name_types = c("surname"),
  name_race_tables = list(surname = wru::surnames2010),
  census.data = NULL,
  key_method = "mutual.inf",
  ...,
  control = NULL
)
}
\arguments{
\item{voter.file}{An object of class data.frame. Must contain a row for each individual being predicted, 
as well as fields named "surname", and/or"first", and/or "middle", containing each 
individual's corresponding names. It must also contain a field named state, which 
contains the two-character lower-case abbreviation for each individual's state of residence (e.g., "nj" for New Jersey).
If using Census geographic data in race/ethnicity
predictions, voter.file must also contain at least one of the following fields: county, tract, block,
and/or place. These fields should contain character strings matching U.S. Census categories. County is three
characters (e.g., "031" not "31"), tract is six characters, and block is four characters. Place is
five characters. See below for other optional fields.}

\item{name_types}{Character vector. Must have elements in "surname", "first", and "middle". Defaults to "surname".}

\item{name_race_tables}{Named list, with as many elements as there are names in \code{name_types}, and names matching
elements in \code{name_types}. Each list element should be a data.frame of unique names (first column) by race (remaining columns),
with conditional probabilities p(Race|Name). Defaults to 
\code{list(surname = wru::surnames2010)}.}

\item{census.data}{A list indexed by two-letter state abbreviations, which contains pre-saved Census geographic data. 
Can be generated using \code{census_geo_api_joint} function. Defaults to \code{NULL}, which 
calls that function.}

\item{key_method}{Method for extracting keynames from name race tables. Currently, only "mutual.inf" is implemented.}

\item{...}{Arguments passed to \code{predict_race}.}

\item{control}{List of control arguments, including 
\itemize{
 \item{max_keynames}{Maximum number of names in census table to use as keynames. Defaults to 100.}
 \item{iter}{Number of MCMC iterations. Defaults to 1000.}
 \item{burnin}{Number of iterations discarded as burnin. Defaults to half of \code{iter}.}
 \item{thin}{Thinning interval for MCMC. Defaults to 1.}
 \item{log_post_interval}{Interval for storing the log_posterior. Defaults to 10.}
 \item{beta_prior}{Parameter for symmetric Dirichlet prior over names for each race. Defaults to 5.}
 \item{gamma_prior}{Parameter for Beta prior over keyname/non-keyname mixture components. Defaults to c(5, 5).}
 \item{verbose}{Print progress information. Defaults to \code{TRUE}.}
 \item{seed}{RNG seed. If \code{NULL}, a seed is generated and stored for reproducibility.}
}}
}
\value{
A named list of predicted distributions over races for each
name type.
}
\description{
Estimate a Bayesian mixed-membership, keyword-assisted race classification model for geo-locations
in a voterfile.
}
