% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/find_keynames.R
\name{.find_keynames}
\alias{.find_keynames}
\title{Obtain keynames from census table. Internal function}
\usage{
.find_keynames(names, freq_table, method = "mutual.inf", n_keynames = 100)
}
\arguments{
\item{names}{Character vector of potential keynames}

\item{freq_table}{Matrix with overall name ferquency, by race (columns)}

\item{method}{Character string. Which method should be used to select keynames?
Currently, the only implemented method is \code{mutual.inf}, which
selectes names with highest mutual information by race.}

\item{n_keynames}{Number of keynames to select by race. Defaults to 100.}
}
\value{
A character matrix of top \code{n_keynames} by race.
}
\description{
Obtain keynames from census table. Internal function
}
\keyword{internal}
