% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_census_data.R
\name{get_census_data}
\alias{get_census_data}
\title{Multilevel Census data download function.}
\usage{
get_census_data(key, states, age = FALSE, sex = FALSE,
  census.geo = "block", retry = 0)
}
\arguments{
\item{key}{A required character object containing a valid Census API key, 
which can be requested \href{https://api.census.gov/data/key_signup.html}{here}.}

\item{states}{which states to extract Census data for, e.g., \code{c("NJ", "NY")}.}

\item{age}{A \code{TRUE}/\code{FALSE} object indicating whether to condition on 
age or not. If \code{FALSE} (default), function will return Pr(Geolocation | Race).
If \code{TRUE}, function will return Pr(Geolocation, Age | Race). 
If \code{\var{sex}} is also \code{TRUE}, function will return Pr(Geolocation, Age, Sex | Race).}

\item{sex}{A \code{TRUE}/\code{FALSE} object indicating whether to condition on 
sex or not. If \code{FALSE} (default), function will return Pr(Geolocation | Race). 
If \code{TRUE}, function will return Pr(Geolocation, Sex | Race). 
If \code{\var{age}} is also \code{TRUE}, function will return Pr(Geolocation, Age, Sex | Race).}

\item{census.geo}{An optional character vector specifying what level of 
geography to use to merge in U.S. Census 2010 geographic data. Currently
\code{"county"}, \code{"tract"}, or \code{"block"} are supported.}

\item{retry}{The number of retries at the census website if network interruption occurs.}
}
\value{
Output will be an object of class \code{list} indexed by state. 
Output will contain the following elements: \code{state}, \code{age}, \code{sex}, 
\code{county}, \code{tract} and \code{block}.
}
\description{
\code{get_census_data} returns county-, tract-, and block-level Census data 
for specified state(s). Using this function to download Census data in advance 
can save considerable time when running \code{predict_race} and \code{census_helper}.
}
\examples{
\dontrun{get_census_data(key = "...", states = c("NJ", "NY"), age = TRUE, sex = FALSE)}
}

