% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getCensusApi2.R
\name{getCensusApi2}
\alias{getCensusApi2}
\title{Census API URL assembler.}
\usage{
getCensusApi2(data_url, key, get, region)
}
\arguments{
\item{data_url}{URL root of the API, including the '?', 
e.g., "http://api.census.gov/data/2010/sf1?".}

\item{key}{A required character object containing user's Census API key, 
which can be requested at http://api.census.gov/data/key_signup.html.}

\item{get}{A character vector of variables to get, 
e.g., c("P0050003","P0050004","P0050005", "P0050006").
If there are more than 50 variables, then function will automatically 
split variables into separate queries.}

\item{region}{Character object specifying which region to obtain data for.
Must contain "for" and possibly "in", 
e.g., "for=block:1213&in=state:47+county:015+tract:*"}
}
\value{
If successful, output will be an object of class \code{data.frame}. 
 If unsuccessful, function prints the URL query that was constructed.
}
\description{
\code{getCensusApi2} assembles URL components for \code{getCensusApi}.
}
\details{
This function assembles the URL components and sends the request to the Census server. 
It is used by the \code{getCensusApi} function. The user should not need to call this 
function directly.
}
\examples{
\dontrun{getCensusApi2(data_url = "http://api.census.gov/data/2010/sf1?", key = "...", get = c("P0050003","P0050004","P0050005", "P0050006"), region = "for=county:*&in=state:34")}

}
\references{
Based on code authored by Nicholas Nagle, which is available at: 
http://rstudio-pubs-static.s3.amazonaws.com/19337_2e7f827190514c569ea136db788ce850.html.
}

