% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pid.R
\docType{data}
\name{pid}
\alias{pid}
\title{Party and race}
\format{A data frame with 3 rows and 7 variables:
\describe{
  \item{party}{Dem/Rep/Ind}
  \item{PID}{0/1/2}
  \item{r_pid_whi}{Pr(PID | White)}
  \item{r_pid_bla}{Pr(PID | Black)}
  \item{r_pid_his}{Pr(PID | Hispanic/Latino)}
  \item{r_pid_asi}{Pr(PID | Asian/Pacific Islander)}
  \item{r_pid_oth}{Pr(PID | Other)}
  #' }}
\usage{
pid
}
\description{
A table for probability of party by race.
}
\examples{
data(pid)
}
\keyword{datasets}
