% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/race.pred.R
\name{race.pred}
\alias{race.pred}
\title{Race prediction function.}
\usage{
race.pred(voters, vars, races, UScensus2010)
}
\arguments{
\item{voters}{An object of class \code{data.frame}. Must contain fields for 
surname ('surname') and geolocation ('tract' or 'block'). 
Optional fields include: party registration ('party'), 'age', and 'sex'.}

\item{vars}{A character vector specifying variables on which to condition 
race predictions. May include "surname", "tract" or "block", and "party".}

\item{races}{A character vector specifying which races to generate 
predictions for.}

\item{UScensus2010}{A TRUE/FALSE object. If TRUE, function will call
\code{census.helper} to merge in U.S. Census 2010 data. If FALSE, 
voters must contain fields specifying Pr(X | Race), where X is the 
set of variables with which to update the prior Pr(Race | Surname). 
These fields may include any of the following: 'r_whi', 'r_bla', 
'r_his', 'r_asi', and/or 'r_oth'.}
}
\value{
Output will be an object of class \code{data.frame}. It will 
 consist of the original user-input data with additional columns that 
 contain predicted probabilities for each race in \code{races}.
}
\description{
\code{race.pred} makes probabilistic estimates of individual-level race/ethnicity.
}
\details{
This function implements the Bayesian race prediction methods outlined in 
Imai and Khanna (2015). The function estimates individual-level 
race/ethnicity, based on surname, geolocation, and party registration.
}
\examples{
race.pred(voters = ..., vars = c("surname", "block"), races = c("asian"), UScensus2010 = FALSE)
race.pred(voters = ..., vars = c("surname", "tract"), races = c("white", "black", "latino"), UScensus2010 = TRUE)
race.pred(voters = ..., vars = c("surname", "tract", "party"), races = c("white", "black", "latino", "asian", "other"), UScensus2010 = TRUE)

}

