% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/voters.R
\docType{data}
\name{voters}
\alias{voters}
\title{Example voter file.}
\format{
A data frame with 10 rows and 12 variables:
\describe{
  \item{VoterID}{Voter identifier (numeric)}
  \item{surname}{Surname}
  \item{state}{State of residence}
  \item{CD}{Congressional district}
  \item{county}{Census county (three-digit code)}
  \item{tract}{Census tract (six-digit code)}
  \item{block}{Census block (four-digit code)}
  \item{precinct}{Voting precinct}
  \item{place}{Voting place}
  \item{age}{Age in years}
  \item{sex}{0=male, 1=female}
  \item{party}{Party registration (character)}
  \item{PID}{Party registration (numeric)}
  #' }
}
\usage{
voters
}
\description{
An example dataset containing voter file information.
}
\examples{
data(voters)
}
\keyword{datasets}
