% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/merge_names.R
\name{wru_data_preflight}
\alias{wru_data_preflight}
\title{Preflight for name data}
\usage{
wru_data_preflight()
}
\description{
Checks if namedata is available in the current working directory, if not
downloads it from github using piggyback. By default, wru will download the
data to a temporary directory that lasts as long as your session does.
However, you may wish to set the \code{wru_data_wd} option to save the 
downloaded data to your current working directory for more permanence.
}
