% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/merge_names.R
\name{wru_data_preflight}
\alias{wru_data_preflight}
\title{Preflight for name data}
\usage{
wru_data_preflight()
}
\description{
Checks if namedata is avialable in the current working directory, if not
downloads it from github using piggyback.
}
